refactor(account-service): move logger to utils from server and changed relevent imports
