+++ apps/account-service/src/modules/auth/auth.service.ts
@@ -1,8 +1,8 @@
 import type { registerSchemaDataType } from './auth.validation.js';
 import { prisma } from '../../core/prisma-client.js';
-import { logger } from '../../core/server.js';
 import { hashPassword } from '../../shared/utils/password.js';
 import { ApiError } from '../../core/api-error.js';
+import { logger } from '../../shared/utils/logger.js';
 
 const registerUser = async (userData: registerSchemaDataType) => {
   const existingUser = await prisma.user.findUnique({
