+++ apps/account-service/src/shared/utils/logger.ts
@@ -1,0 +1,3 @@
+import { createLogger } from '@repo/logger';
+
+export const logger = createLogger({ service: 'account-service' });
